requirements.txt
python_telegram_bot==20.1
pytz~=2022.1
spacy~=3.5.2
geopy~=2.3.0
timezonefinder~=6.2.0